% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_last_Sat}
\alias{get_last_Sat}
\title{get the most recent Sunday}
\usage{
get_last_Sat(d = Sys.Date())
}
\arguments{
\item{d}{the day, as class `Date`}
}
\value{
the previous Sunday with class `Date`
}
\description{
get the most recent Sunday
}
