% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Sat}
\alias{get_Sat}
\title{get the previous Saturday}
\usage{
get_Sat(d = Sys.Date(), prev = T)
}
\arguments{
\item{d}{the day, as class `Date`}

\item{prev}{if set to `TRUE` it find the previous Saturday, `FALSE` finds the next}
}
\value{
a Saturday with class `Date`
}
\description{
get the previous Saturday
}
