% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_Sun}
\alias{get_Sun}
\title{get the next Sunday}
\usage{
get_Sun(d = Sys.Date())
}
\arguments{
\item{d}{the day, as class `Date`}
}
\value{
the next Sunday with class `Date`
}
\description{
get the next Sunday
}
