% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl_helpers.R
\name{get_toggl_day}
\alias{get_toggl_day}
\title{Get Toggl time for a single day}
\usage{
get_toggl_day(date, toggl_token, min = 5)
}
\arguments{
\item{date}{the day to get the time for}

\item{toggl_token}{the Toggl token of interest}

\item{min}{what to round the minues to, default 5}
}
\value{
a dataframe will the user's times for the day
}
\description{
Get Toggl time for a single day
}
