% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl_helpers.R
\name{get_toggl_entries}
\alias{get_toggl_entries}
\title{Get weekly Toggl entries since date}
\usage{
get_toggl_entries(date = Sys.Date(), toggl_token)
}
\arguments{
\item{date}{the date to get the Toggl entries from. Should be class `Date`.}

\item{toggl_token}{the user's toggl token}
}
\value{
a data frame with billed hours.
}
\description{
This method pull Toggl entries from the date provided to the
preceeding Sunday and then format them into a clean dataframe. Make sure
that your Toggl projects follow the Toggl project guidelines (see below)
}
\details{
Make sure to follow the Toggl project format guidelines or this
function will not work. The structure is as below:

For single bill codes:
"Client Task 123456:001"
or
"Client Task 123456:001:002:003"
for multiple tasks, time is split evenly

For multiple bill codes use the following structure:
"Client Task 123456:001&987654:002"
or
"Client Task 123456:001&987654:002 80/20"
or
"Client Task 123456:001&987654:002&345678:003 60/20/20"

Ratios are optionally specified integers (no decimals) and do not need to sum 100.
}
