% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{NAVlogin}
\alias{NAVlogin}
\title{Login to insideNCI}
\usage{
NAVlogin(un, pass)
}
\arguments{
\item{un}{the user's Navigant username}

\item{pass}{the user's Navigant password}
}
\value{
an rvest html session logged into insidenNCI
}
\description{
Login to insideNCI
}
