% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheet_helpers.R
\name{ts_set_rows}
\alias{ts_set_rows}
\title{Set # of rows in a timesheet}
\usage{
ts_set_rows(session, n)
}
\arguments{
\item{session}{the rvest html session at the timesheet page}

\item{n}{the number of rows needed}
}
\value{
an rvest html session
}
\description{
Set # of rows in a timesheet
}
