% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{report_create}
\alias{report_create}
\title{Create and save a report from Toggl}
\usage{
report_create(user, pass, period_end_date = get_Sat())
}
\arguments{
\item{pass}{the user's Navigant password}

\item{period_end_date}{the period end date to make the timesheet for, defaults
for the previous Saturday but can run the next saturday easily by setting
this value to `get_Sat(prev = F)`}

\item{un}{the user's Navigant username}
}
\value{
the resultant html session, can be parsed for errors
}
\description{
Create and save a report from Toggl
}
